implements patches

+++ compiler/src/lsp/lsp.hpp
@@ -4,5 +4,6 @@
 
 auto lsp_diagnostics(const std::string &src) -> std::string;
 auto lsp_tokens(const std::string &src) -> std::string;
-auto lsp_completions(const std::string &src, int line, int col) -> std::string;
-auto lsp_hover(const std::string &src, int line, int col) -> std::string;
+auto lsp_completions(const std::string &src, size_t line, size_t col)
+    -> std::string;
+auto lsp_hover(const std::string &src, size_t line, size_t col) -> std::string;
